requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
hyperliquid-python-sdk==0.20.1
eth-account==0.13.4
eth-utils==5.1.0
python-dotenv==1.0.1
requests==2.32.5
slowapi==0.1.9

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1